% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_bim.R
\name{sim_bim}
\alias{sim_bim}
\title{Construct a variant table in plink BIM format with several chromosomes.}
\usage{
sim_bim(m_loci = 1e+05, n_chr = 22, pos_gap = 1000)
}
\arguments{
\item{m_loci}{The number of loci.}

\item{n_chr}{Number of chromosomes to simulate.
Chromosome assignments are not biologically meaningful, as all loci are drawn independently (no LD).}

\item{pos_gap}{The number of basepairs to space variants.}
}
\value{
A \code{tibble} with these columns:
\itemize{
\item \code{chr}: The constructed chromosome positions.
\item \code{id}: Locus ID values (autocompleted by \code{\link[genio:make_bim]{genio::make_bim()}} to match row numbers).
\item \code{posg}: Genetic distance (autocompleted by \code{\link[genio:make_bim]{genio::make_bim()}} to \code{0}, all missing).
\item \code{pos}: locus positions, starting at \code{pos_gap} for each chromosome and increasing in increments of \code{pos_gap}.
\item \code{ref}: Reference allele (autocompleted by \code{\link[genio:make_bim]{genio::make_bim()}} to \code{1}).
\item \code{alt}: Alternative allele (autocompleted by \code{\link[genio:make_bim]{genio::make_bim()}} to \code{2}).
}
}
\description{
Purpose is to build an artificial BIM table with several chromosomes, which some methods require, such as leave-one-chromosome-out (LOCO) estimation.
This function produces chromosomes with approximately equal numbers of variants.
Positions are spaced uniformly by a given size.
}
\examples{
# simulate a small table
m_loci <- 100
bim <- sim_bim( m_loci )

}
