% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_geno.R
\name{sim_geno}
\alias{sim_geno}
\title{Simulate genotypes from an admixed family structure}
\usage{
sim_geno(
  admix_proportions_1,
  inbr_subpops,
  fam = NULL,
  ids = NULL,
  m_loci = 1e+05,
  p_anc = NULL,
  verbose = TRUE
)
}
\arguments{
\item{admix_proportions_1}{The admixture proportions matrix of the founders (required to draw the founder genotypes from the admixture model).}

\item{inbr_subpops}{The vector of intermediate subpopulation inbreeding/FST values for admixture model.}

\item{fam}{The pedigree structure as a plink FAM table, or \code{NULL} if no family structure is to be simulated.}

\item{ids}{The list of IDs from \code{fam} above split into non-overlapping generations.
Ignored if \code{fam} is \code{NULL}.}

\item{m_loci}{The number of loci.}

\item{p_anc}{The desired ancestral allele frequencies (scalar or length-\code{m_loci} vector passed to \code{\link[bnpsd:draw_all_admix]{bnpsd::draw_all_admix()}}).
By default (\code{NULL}), ancestral allele frequencies are drawn randomly.}

\item{verbose}{If \code{TRUE} reports progress, otherwise it is silent.}
}
\value{
A list containing the following elements:
\itemize{
\item \code{X}: The simulated genotype matrix.
\item \code{p_anc}: The vector of ancestral allele frequencies (required to simulate traits with desired heritability).
}
}
\description{
This function draws genotypes from an admixture model and optionally propagates them through a pedigree.
The admixture genotypes are calculated using \code{\link[bnpsd:draw_all_admix]{bnpsd::draw_all_admix()}} and simply returned if \code{fam} is \code{NULL}.
If \code{fam} is not \code{NULL}, genotypes are propagated through the pedigree using \code{\link[simfam:geno_last_gen]{simfam::geno_last_gen()}}.
The admixture model by default redraws fixed loci from the model.
However, loci may become fixed after propagation through the pedigree.
This function identifies these fixed loci (using \code{\link[bnpsd:fixed_loci]{bnpsd::fixed_loci()}}) and redraws them first from the admixture model and then through the pedigree, iterating until no loci are fixed.
}
\examples{
# draw population parameters using `sim_pop` first
# a small example
data <- sim_pop( n_ind = 50, G = 3 )

# now draw genotypes!
m_loci <- 100
data2 <- sim_geno(
    data$admix_proportions_1,
    data$inbr_subpops,
    data$fam,
    data$ids,
    m_loci
)

# genotype matrix
data2$X

# ancestral allele frequencies
data2$p_anc

}
