% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_pop.R
\name{sim_pop}
\alias{sim_pop}
\title{Simulate population structure}
\usage{
sim_pop(
  n_ind = 1000,
  k_subpops = 3,
  bias_coeff = 0.5,
  fst = 0.3,
  G = 1,
  verbose = TRUE
)
}
\arguments{
\item{n_ind}{The number of individuals.}

\item{k_subpops}{The number of intermediate subpopulations for admixture model.}

\item{bias_coeff}{The bias coefficient of the admixed individuals (for the founders if \code{G > 1}).}

\item{fst}{The FST of the admixed individuals (for the founders if \code{G > 1}).}

\item{G}{The number of generations for random family.
The \code{G = 1} case is no family structure (just admixture).
For \code{G > 1} first admixed founders are simulated, then the family structure is simulated from them.}

\item{verbose}{If \code{TRUE} reports progress, otherwise it is silent.}
}
\value{
A list containing the following elements:
\itemize{
\item \code{admix_proportions_1}: The true admixture proportions of the founders (required to draw genotypes for admixed founders and seed the family structure if \code{G > 1})
\item \code{admix_proportions}: The true admixture proportions of the joint admixed family model (final generation only).
\item \code{inbr_subpops}: The vector of intermediate subpopulation inbreeding/FST values for admixture model.
\item \code{fam}: If \code{G > 1}, the pedigree structure as a plink FAM table, pruned with \code{\link[simfam:prune_fam]{simfam::prune_fam()}} to include only ancestors of the final generation; otherwise \code{NULL}.
\item \code{ids}: If \code{G > 1}, a list of IDs from \code{fam} above split into non-overlapping generations; otherwise \code{NULL}.
\item \code{kinship}: The true kinship of the joint admixed family model (final generation only).
}
}
\description{
This function defines admixture and optionally family parameters establishing how individuals are related to each other.
Admixture proportions are deterministically constructed from the linear 1D admixture model \code{\link[bnpsd:admix_prop_1d_linear]{bnpsd::admix_prop_1d_linear()}}.
If \code{fst < 0.5} the intermediate inbreeding coefficients are proportional to their index (\code{1:k_subpops}), otherwise they are proportional to \code{(1:k_subpops) + 1}.
If \code{G > 1}, a random family structure is drawn using \code{\link[simfam:sim_pedigree]{simfam::sim_pedigree()}}.
No per-locus parameters are constructed or simulated with this function.
}
\examples{
# a small example
data_pop <- sim_pop( n_ind = 50, G = 3 )

# parameters for admixture model of founders
data_pop$admix_proportions_1
data_pop$inbr_subpops

# parameters of family structure
data_pop$fam
data_pop$ids

# parameters for final generation
data_pop$admix_proportions
data_pop$kinship

}
\seealso{
The \code{bnpsd} and \code{simfam} packages.
}
