% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_trait_env.R
\name{sim_trait_env}
\alias{sim_trait_env}
\title{Simulate a complex trait}
\usage{
sim_trait_env(
  X,
  p_anc,
  m_causal,
  herit,
  env = NA,
  env_var = 1,
  k_subpops = NA,
  fes = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{X}{The genotype matrix.}

\item{p_anc}{The vector of ancestral allele frequencies.}

\item{m_causal}{The number of causal loci for the trait, selected randomly from among the simulated loci.}

\item{herit}{The trait heritability.}

\item{env}{A string describing environment model.
Only \code{NA} (no environment) or "gcta" are accepted.}

\item{env_var}{The variance of the environment effect.
Ignored if \code{env} is \code{NA}.}

\item{k_subpops}{The number of intermediate subpopulations for admixture model.
Used for \code{env = 'gcat'} only.}

\item{fes}{If \code{FALSE} (default) constructs trait from Random Coefficients (RC) model.
If \code{TRUE}, the Fixed Effect Sizes (FES) model is used instead.}

\item{verbose}{If \code{TRUE} reports progress, otherwise it is silent.}
}
\value{
A list containing the following elements:
\itemize{
\item \code{trait}: The simulated trait vector.
\item \code{causal_indexes}: The vector of randomly selected causal loci indexes.
\item \code{causal_coeffs}: The vector of simulated regression coefficients for causal loci.
}
}
\description{
Mostly a wrapper around \code{\link[simtrait:sim_trait]{simtrait::sim_trait()}}, but can optionally add a structured environment effect.
However, in that case the heritability is no longer as specified.
}
\examples{
# draw population parameters using `sim_pop` first
# a small example
k_subpops <- 3
data <- sim_pop( n_ind = 50, G = 3, k_subpops = k_subpops )

# then draw genotypes
m_loci <- 100
data2 <- sim_geno(
    data$admix_proportions_1,
    data$inbr_subpops,
    data$fam,
    data$ids
)

# now draw trait!
m_causal <- 5
herit <- 0.5
data3 <- sim_trait_env(
    data2$X,
    data2$p_anc,
    m_causal,
    herit,
    env = 'gcat',
    k_subpops = k_subpops
)

# the simulated trait
data3$trait
# indexes of randomly-selected causal loci
data3$causal_indexes
# coefficients of causal loci
data3$causal_coeffs

}
